+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Improvements
 - Improved the fio input plugins to allow more generic metadata definitions.
 - Modifed the way custom plugin widgets are handles and create a base class
   for custom widgets to inherit from.
+- Added a generic xscale to fio plugins to have a more generalized use case covered.
 
 Bugfixes
 --------